fix type hinting: maxFeeRate -> float

+++ defichain/transactions/remotedata/node.py
@@ -24,12 +24,12 @@ class RemoteDataNode(RemoteData):
             unspent.append({"txid": txid, "vout": index, "value": value, "scriptPubKey": script})
         return unspent
 
-    def test_tx(self, hex: str, maxFeeRate: int = None) -> bool:
+    def test_tx(self, hex: str, maxFeeRate: float = None) -> bool:
         try:
             self.node.rawtransactions.testmempoolaccept([hex], maxFeeRate)
             return True
         except:
             return False
 
-    def send_tx(self, hex: str, maxFeeRate: int = None) -> str:
+    def send_tx(self, hex: str, maxFeeRate: float = None) -> str:
         return self.node.rawtransactions.sendrawtransaction(hex, maxFeeRate)
